feat: Add jQuery script to fetch greeting in French
The code changes include adding a new HTML file `9-main.html` and a new JavaScript file `9-script.js`. The HTML file includes a jQuery script tag to load the jQuery library and the `9-script.js` file. The JavaScript file contains a jQuery function that makes an AJAX GET request to `https://hellosalut.stefanbohacek.dev/?lang=fr` and updates the text of the `hello` div with the response.

Based on the recent repository commits, it seems that the changes made in this commit are related to fixing a bug or implementing a new feature.